Fix issue with disappearing about me section
The main text of the about me section kept disappearing shortly loading
the index page. For now, this can be fixed by using the internal content
of these markdown files instead of the generated HTML. Since the content
is a single paragraph this does not make a difference for constructing
the page.